[#96] Refactor: Update package structure and configuration for chat room services
