+++ src/main/java/soon/capstone/domain/chatroom/service/ChatRoomTeamMemberService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import soon.capstone.domain.chatroom.entity.ChatRoom;
 import soon.capstone.domain.chatroom.entity.ChatRoomTeamMember;
-import soon.capstone.domain.chatroom.repository.ChatRoomRepository;
-import soon.capstone.domain.chatroom.repository.ChatRoomTeamMemberRepository;
+import soon.capstone.domain.chatroom.repository.chatroom.ChatRoomRepository;
+import soon.capstone.domain.chatroom.repository.member.ChatRoomTeamMemberRepository;
 import soon.capstone.domain.chatroom.service.dto.request.ChatRoomAddMemberServiceRequest;
 import soon.capstone.domain.teammember.entity.TeamMember;
 import soon.capstone.domain.teammember.repository.TeamMemberRepository;
