+++ src/test/java/soon/capstone/domain/chatroom/service/ChatRoomServiceTest.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import soon.capstone.IntegrationTestSupport;
 import soon.capstone.domain.chatroom.entity.ChatRoom;
 import soon.capstone.domain.chatroom.entity.ChatRoomTeamMember;
-import soon.capstone.domain.chatroom.repository.ChatRoomRepository;
-import soon.capstone.domain.chatroom.repository.ChatRoomTeamMemberRepository;
+import soon.capstone.domain.chatroom.repository.chatroom.ChatRoomRepository;
+import soon.capstone.domain.chatroom.repository.member.ChatRoomTeamMemberRepository;
 import soon.capstone.domain.chatroom.service.dto.request.ChatRoomCreateServiceRequest;
 import soon.capstone.domain.chatroom.service.dto.request.ChatRoomFinishServiceRequest;
 import soon.capstone.domain.chatroom.service.dto.request.ChatRoomResumeServiceRequest;
